02.05 라이크 추가, 삭제 가능. 비회원 좋아요 기능 처리. 비동기식처리 작업중

+++ src/main/java/zerogreen/eco/service/detail/LikesService.java
@@ -1,6 +1,5 @@
 package zerogreen.eco.service.detail;
 
-import zerogreen.eco.dto.detail.LikesDto;
 import zerogreen.eco.entity.userentity.BasicUser;
 
 public interface LikesService {
@@ -11,9 +10,6 @@ public interface LikesService {
     //나 얼만큼 좋아해요
     Long cntLikes(Long sno);
     //개별 라이크
-    int cntMemberLike(Long sno, Long mno);
-    //라이크 데이터
-    //LikesDto liking(Long id, String username);
-
+    Long cntMemberLike(Long sno, Long mno);
 
     }
